Improve editable keybinding settings rows
